Renamed dashboard page to courses
