+++ src/components/courses/CourseCard.tsx
