+++ src/styles/globals.css
@@ -88,7 +88,7 @@
     @apply flex flex-col justify-center items-center gap-8 p-8 py-10 sm:p-20 w-full min-h-full;
   }
 
-  .container-dashboard {
+  .container-courses {
     @apply flex flex-col items-center gap-8 p-8 py-10 w-full min-h-full justify-start md:h-[calc(100vh_-_64px)];
   }
